+++ lib/db_core/migrations/index.dart
@@ -1,0 +1,2 @@
+export 'main_store_migration_runner.dart';
+export 'main_store_migration_types.dart';